+++ CHANGES.rst
@@ -1,12 +1,16 @@
-Changes in Matrix iOS SDK in 0.xx.xx (2019-xx-xx)
+Changes in Matrix iOS SDK in 0.15.1 (2019-xx-xx)
+===============================================
+Improvements:
+ * Well-known: Expose "m.integrations" according to [MSC1957](https://github.com/matrix-org/matrix-doc/pull/1957) (vector-im/riot-ios#2815).
+ * MXSession: Expose and store homeserverWellknown.
+
+Changes in Matrix iOS SDK in 0.15.0 (2019-11-06)
 ===============================================
 
 Improvements:
  * MX3PidAddManager: Add User-Interactive Auth to /account/3pid/add (vector-im/riot-ios#2744).
  * MXSession: On resume, make the first /sync request trigger earlier (vector-im/riot-ios#2793).
  * MXCrypto: Do not fail to decrypt when there is nothing to decrypt (redacted events).
- * Well-known: Expose "m.integrations" according to [MSC1957](https://github.com/matrix-org/matrix-doc/pull/1957) (vector-im/riot-ios#2815).
-* MXSession: Expose and store homeserverWellknown.
 
 Bug fix:
  * Room members who left are listed with the actual members (vector-im/riot-ios#2737).
@@ -14,6 +18,9 @@ Bug fix:
  * MXHTTPOperation: Make urlResponseFromError return the url response in case of MXError.
  * MXHTTPOperation: Fix a crash in `-mutateTo:` method when operation parameter is nil.
  * VoIP: Fix regression when using a TURN server (vector-im/riot-ios#2796).
+
+API break:
+ * MXBackgroundModeHandler: Update interface and now use a single method that return a MXBackgroundTask.
 
 Changes in Matrix iOS SDK in 0.14.0 (2019-10-11)
 ===============================================